requirements.txt
Flask==3.1.1
Flask-Cors==6.0.0
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
onnxruntime==1.22.0
faiss-cpu==1.11.0
numpy==2.2.2
opencv-python==4.11.0.86
Pillow==11.1.0
scikit-learn==1.6.1
python-dotenv==1.1.0
SQLAlchemy==2.0.41
Mako==1.3.10
alembic==1.16.1
PyMySQL==1.1.1
pyngrok==7.2.11